requirements.txt
argparse
beautifulsoup4
gitpython
google-api-python-client
jinja2
mailbox
requests
scholarly
tabulate
tqdm
